Added insufficient material
GUI changes:
tweaked elimination chess promotion options
fixed the eval bar
added option for randomly generated boards to not have a king

Engine changes:
Minor speedups (1-2%)
Insufficient material: No elo change
TT Move: 60 elo
Change quiescence depth to 5: 30 elo
Soft TM: 10 elo
Null move pruning: 15 elo

+++ tester/src/lib.rs
@@ -8,7 +8,7 @@ use liberty_chess::positions::{
 };
 
 /// The test positions for the match
-pub const POSITIONS: &[(&str, &str, u16, f64)] = &[
+pub const POSITIONS: &[(&str, &str, u32, f64)] = &[
   ("startpos", STARTPOS, 20, 0.4395),
   ("rectangle", CAPABLANCA_RECTANGLE, 20, 0.3343),
   ("capablanca", CAPABLANCA, 25, 0.3846),
